pyproject.toml
[tool.poetry]
name = "discord-cftools-wfr"
version = "0.1.0"
description = ""
authors = ["Vebnik(Nik) <vebnik111@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.1"
requests = "^2.28.2"
pydantic = "^1.10.4"
pytest = "^7.2.1"
tortoise-orm = "^0.19.3"
aerich = "^0.7.1"

[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start='main:main'
migrate='src.database.migrate:migrate'